Notify permitted players when we encounter alt accounts; Fixes #133
This commit adds a new AltAccountNotifer which allows to schedule
an async check for a player. The check gets all players with that IP
and checks them for last login. If any have logged in during the
last two weeks, a notification is sent to all players with the
"xlogin.altinfo" permission.