Reduce /xlol spam, Improve time-IP-based limiting, Add Unit Tests
This commit reduces spam sent by the rate limit manager. It does so
by only issuing one message for all infringing IPs.

Additionally, it completely reworks the IP rate limiting to use a
more object-oriented approach similar to the other limits. This
allows to keep times per IP and as a result, join counts are now
always displayed correctly, since they are no longer abused for time
limits.

Furthermore, it improves the /xlol command to use the new system and
be more object-oriented in general.

It also adds Unit Tests for SimpleRateLimit and IpRateLimit.